server/requirements.txt
fastapi
numpy
python-multipart
uvicorn
noisereduce
soundfile
openai-whisper
python-dotenv
boto3
whisperx
speechlib
googletrans